% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_c.R
\name{f_to_c}
\alias{f_to_c}
\title{convert fahrenheit to celcius}
\usage{
f_to_c(fahrenheit)
}
\arguments{
\item{fahrenheit}{- number, temperature in fahrenheit}
}
\value{
number, temperature in celcius
}
\description{
convert fahrenheit to celcius
}
